Prevent ambiguity when parsing all job names, by excluding the delimiter from the unique part of the 'baseJobName'.
This prevents confusion when using branch names that have dashes in them, from being mistaken as a job template.
e.g. 'MyProject-RunTests-fix-something-on-master' would be interpreted as a new job template, when it was actually
produced from the 'MyProject-RunTests-master' job template and the branch 'fix-something-on-master'.
The confusion would result in lots of invalid jobs being created for the 'new' job template.